feat: Get CLI running on windows (#27)
Per strands-agents/tools#17 some tools do not work on windows and they fail on importing the tools. To workaround that, conditionally include tools when running on windows.

Also included an update to the strand tool to make all tools available since we now have centralized logic for that.

Also add workflows to run tests on windows + MacOS

---------

Co-authored-by: Mackenzie Zastrow <zastrowm@users.noreply.github.com>